lens: move write-cache `get`/`list` into `fstree` section
Write-cache is nothing more but another FSTree now, so move lens commands `get`
and `list` to `fstree` section. Delete `write-cache` section, as it no longer
contains any commands.

Closes #3824.
