+++ cmd/neofs-lens/internal/fstree/get.go
@@ -1,4 +1,4 @@
-package writecache
+package fstree
 
 import (
 	"fmt"
@@ -11,7 +11,7 @@ import (
 var getCMD = &cobra.Command{
 	Use:   "get",
 	Short: "Object inspection",
-	Long:  `Get specific object from a write-cache.`,
+	Long:  `Get specific object from FSTree.`,
 	Args:  cobra.NoArgs,
 	RunE:  getFunc,
 }
@@ -21,21 +21,22 @@ func init() {
 	common.AddComponentPathFlag(getCMD, &vPath)
 	common.AddOutputFileFlag(getCMD, &vOut)
 	common.AddPayloadOnlyFlag(getCMD, &vPayloadOnly)
+	AddDepthFlag(getCMD, &vDepth)
 }
 
 func getFunc(cmd *cobra.Command, _ []string) error {
-	wc, err := openWC()
+	fst, err := openFSTree()
 	if err != nil {
 		return err
 	}
-	defer wc.Close()
+	defer fst.Close()
 
 	addr, err := oid.DecodeAddressString(vAddress)
 	if err != nil {
 		return fmt.Errorf("invalid address: %w", err)
 	}
 
-	obj, err := wc.Get(addr)
+	obj, err := fst.Get(addr)
 	if err != nil {
 		return fmt.Errorf("could not fetch object: %w", err)
 	}
