+++ cmd/neofs-lens/root.go
@@ -8,7 +8,6 @@ import (
 	"github.com/nspcc-dev/neofs-node/cmd/neofs-lens/internal/meta"
 	"github.com/nspcc-dev/neofs-node/cmd/neofs-lens/internal/object"
 	"github.com/nspcc-dev/neofs-node/cmd/neofs-lens/internal/storage"
-	"github.com/nspcc-dev/neofs-node/cmd/neofs-lens/internal/writecache"
 	"github.com/nspcc-dev/neofs-node/misc"
 	"github.com/nspcc-dev/neofs-node/pkg/util/gendoc"
 	"github.com/spf13/cobra"
@@ -40,7 +39,6 @@ func init() {
 	command.Flags().Bool("version", false, "Application version")
 	command.AddCommand(
 		meta.Root,
-		writecache.Root,
 		storage.Root,
 		object.Root,
 		fstree.Root,
